% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{run_example}
\alias{run_example}
\title{Run the application}
\usage{
run_example(display.mode = c("normal", "showcase"))
}
\arguments{
\item{display.mode}{The mode in which to display the application. If \code{"showcase"}, application code is
diplayed along with the output. If \code{"normal"} (default), the code is hidden.}
}
\description{
Run the Shiny Application
}
